+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All Notable changes to `drupal/opening-hours` module.
 
+## [Unreleased]
+
+### Fixed
+
+* DMOH-53: Fixed not unique element wrapper.
+* DMOH-53: Fixed detecting if the element was submitted.
+
 ## [8.x-1.0]
 
 ### Added